Bank Account samples with console application
This console application consist of 8 classes. One of them is parent class which is bank account which can points other bank account types to generate rules according to types. It is an example of bank account types with their rules and can be visualized better with other languages or can be used to develop account application.